Move app.go content to root.go
